Refactor app layout into dedicated components
